se establece ResponseEntity para todas las entidades del proyecto

+++ fuentes/BACK/backend/src/main/java/trainer/api/backend/controller/ObjetivoController.java
@@ -1,5 +1,6 @@
 package trainer.api.backend.controller;
 
+import org.apache.commons.lang3.ObjectUtils;
 import lombok.AllArgsConstructor;
 import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
@@ -19,18 +20,18 @@ public class ObjetivoController {
 
     @PostMapping("objetivo")
     public ResponseEntity<?> create(@RequestBody ObjetivoDTO objetivoDto){
-        Objetivo clienteSave = null;
+        Objetivo clienteSave;
         try {
             clienteSave = objetivoService.save(objetivoDto);
             return new ResponseEntity<>(MensajeResponse.builder()
                     .mensaje("Guardado correctamente")
                     .object(ObjetivoDTO.builder()
-                            .cumplido(objetivoDto.getCumplido())
-                            .descripcion(objetivoDto.getDescripcion())
-                            .fechaFin(objetivoDto.getFechaFin())
-                            .fechaRegistro(objetivoDto.getFechaRegistro())
-                            .id(objetivoDto.getId())
-                            .usuarioId(objetivoDto.getUsuarioId())
+                            .cumplido(clienteSave.getCumplido())
+                            .descripcion(clienteSave.getDescripcion())
+                            .fechaFin(clienteSave.getFechaFin())
+                            .fechaRegistro(clienteSave.getFechaRegistro())
+                            .id(clienteSave.getId())
+                            .usuarioId(clienteSave.getUsuarioId())
                             .build())
                     .build()
                     , HttpStatus.CREATED);
@@ -44,9 +45,9 @@ public class ObjetivoController {
         }
     }
 
-    @PutMapping("objetivo")
+    @PutMapping("objetivo/{id}")
     public ResponseEntity<?> update(@RequestBody ObjetivoDTO objetivoDto, @PathVariable Long id){
-        Objetivo objetivoUpdate = null;
+        Objetivo objetivoUpdate;
         try {
             if (objetivoService.existsById(id)) {
                 objetivoDto.setId(id);
@@ -62,15 +63,14 @@ public class ObjetivoController {
                                 .usuarioId(objetivoUpdate.getUsuarioId())
                                 .build())
                         .build()
-                        , HttpStatus.CREATED);
-            } else {
-                return new ResponseEntity<>(
-                        MensajeResponse.builder()
-                                .mensaje("El registro que intenta actualizar no se encuentra en la base de datos.")
-                                .object(null)
-                                .build()
-                        , HttpStatus.NOT_FOUND);
+                        , HttpStatus.OK);
             }
+            return new ResponseEntity<>(
+                    MensajeResponse.builder()
+                            .mensaje("El registro que intenta actualizar no se encuentra en la base de datos.")
+                            .object(null)
+                            .build()
+                    , HttpStatus.NOT_FOUND);
         } catch (DataAccessException exDt) {
             return new ResponseEntity<>(
                     MensajeResponse.builder()
@@ -85,7 +85,10 @@ public class ObjetivoController {
     public ResponseEntity<?> deleteById(@PathVariable Long id){
         try {
             Objetivo objetivoDelete = objetivoService.findById(id);
-            objetivoService.delete(objetivoDelete);
+            if(ObjectUtils.isNotEmpty(objetivoDelete)){
+                objetivoService.delete(objetivoDelete);
+                return new ResponseEntity<>(objetivoDelete, HttpStatus.OK);
+            }
             return new ResponseEntity<>(objetivoDelete, HttpStatus.NO_CONTENT);
         } catch (DataAccessException exDt) {
             return new ResponseEntity<>(
@@ -93,35 +96,34 @@ public class ObjetivoController {
                             .mensaje(exDt.getMessage())
                             .object(null)
                             .build()
-                    , HttpStatus.METHOD_NOT_ALLOWED);
+                    , HttpStatus.CONFLICT);
         }
     }
 
     @GetMapping("objetivo/{id}")
     public ResponseEntity<?> showById(@PathVariable Long id){
-        Objetivo objetivo = objetivoService.findById(id);
+        var objetivo = objetivoService.findById(id);
 
-        if (objetivo == null) {
+        if (ObjectUtils.isNotEmpty(objetivo)) {
             return new ResponseEntity<>(
-                    MensajeResponse.builder()
-                            .mensaje("El registro que intenta buscar, no existe!!")
-                            .object(null)
-                            .build()
-                    , HttpStatus.NOT_FOUND);
-        }
-
-        return new ResponseEntity<>(
                 MensajeResponse.builder()
-                        .mensaje("")
-                        .object(ObjetivoDTO.builder()
-                                .cumplido(objetivo.getCumplido())
-                                .descripcion(objetivo.getDescripcion())
-                                .fechaFin(objetivo.getFechaFin())
-                                .fechaRegistro(objetivo.getFechaRegistro())
-                                .id(objetivo.getId())
-                                .usuarioId(objetivo.getUsuarioId())
-                                .build())
-                        .build()
+                    .mensaje("Registro rescatado correctamente")
+                    .object(ObjetivoDTO.builder()
+                        .cumplido(objetivo.getCumplido())
+                        .descripcion(objetivo.getDescripcion())
+                        .fechaFin(objetivo.getFechaFin())
+                        .fechaRegistro(objetivo.getFechaRegistro())
+                        .id(objetivo.getId())
+                        .usuarioId(objetivo.getUsuarioId())
+                        .build())
+                    .build()
                 , HttpStatus.OK);
+        }
+        return new ResponseEntity<>(
+            MensajeResponse.builder()
+                .mensaje("El registro que intenta buscar no existe")
+                .object(null)
+                .build()
+            , HttpStatus.NOT_FOUND);
     }
 }
